+++ test/tests/ignore.rb
@@ -298,7 +298,6 @@ IGNORE_JSON = <<JSON
       "note": "Here's a note!"
     }
   ],
-  "updated": "2019-04-02 12:15:05 -0700",
   "brakeman_version": "4.5.0"
 }
 JSON